Dodge roll and Stamina
Got the dodge roll working as good as it can for the time being. Will need a character animation in order to proceed further. Got stamina working with some bugs. While sprinting and dodging stamina depletes. While standing and moving stamina regenerates. If the player mashes the dodge, it causes bugs with the stamina and causes a slew of problems. It is also not intended for the player to be able to spam dodge. Will fix later